Reworking tests with the rspec-puppet-facts gem.
This new gem adds the ability to perform spec testing on all supported operating
systems defined in metadata.json by use of the on_supported_os function.

A consequence of this update the manifests had to be updated to check against
the operatingsystem fact instead of the osfamily fact. Also, all $::kernel
references had to be removed. The manifests are not as dynamic as they were
previously but there is currently no need for that flexibility now.