packetbeat_version: Fixing the fact and testing.
Travis flagged that during testing of the fact the case that was supposed
to return a real value actually returned nil. This was an unexpected side
effect of the rubocop cleanup and full testing was not performed. This
commit restores the testing and the functionality of the fact itself
and its test cases.